stopping point trying to fix styles
